Replace `*-template` files with `*.local` pattern
Previously, this dotfiles setup copied `*-template` files to the $HOME
directory and used them to include shared config files from the
`dotfiles` repo.  The reasoning was that local modifications could be
made _before_ as well as after the shared configuration, but is also
makes it more difficult to manage _shared_ configuration changes and
commit them to the repository.

A more common pattern is symlink'ing to shared config files that include
local modifications from optional `*.local` files.  With this pattern,
changes to the shared configuration are made in the standard locations,
and appear as changes to the linked files in the repository dir where
they can be committed easily.  The rare need to provide additional
local modifications before the shared config can still be addressed in
other ways.