git: Change how shared config is loaded
For most shared config files, it seems to work best if the actual config
file is linked to the shared configuration in source control, and it can
include an optional `*.local` file for the occasional system-specific
settings.  However, some tools (like SourceTree, for example) insist on
automatically re-applying their settings if they aren't already present
in the `.gitconfig` file itself.  This makes it difficult to migrate
local-only settings into an include file and not have them reappear in
the shared config later.

For Git, it may work better to put the local-only settings in the main
`.gitconfig` file instead, and have it include the shared settings.
We can do this by installing an initial template file using rcm's
`COPY_ALWAYS` option, so any further changes to the file won't appear
as modifications in the Git repository.  Any changes to `~/.gitconfig`
that should be shared can be migrated into the `~/.gitconfig.shared`
file manually.