zsh: Change how optional scripts are sourced
The command `[[ -f file ]] && source file` will source an optional file
only if it exists, but it has an unintended side-effect.  The `&&`
operator works much like the corresponding logical operator in many
programming languages, which means that we only get a success status if
_both_ commands succeed.  If the test for the optional file fails then
this command produces a non-zero exit status in `$?`.  If it happens to
be the last statement executed in a script then the script will also
exit with a non-zero status.

To avoid a possible "failure" status from the `.zshrc` script, we need
to reverse the check for optional files using the `||` operator.  The
command `[[ ! -f file ]] || source file` succeeds if _either_ command
succeeds, so we get a success status in `$?` if the file does _not_
exist or if sourcing the file succeeds.