brew: Disable `dropbox` cask for now
As of the commit date, Apple Silicon support for Dropbox is still in
beta, and the `dropbox` cask will download an installer that only
supports Intel.  Since the cask won't do the right thing for M1 Macs,
it's best to just disable the cask and manually install the Dropbox
Beta.  This can be turned back on when full Apple Silicon support is
available in the cask.