init: Don't show full path in macOS 11+ Finder
The Finder UI was redesigned starting in macOS 11 Big Sur.  It no longer
has a standard window title bar at the top.  Instead, the pathname now
shares space with the toolbar, and it also uses a larger font, so there
is significantly less space.  If the pathname doesn't fit then it is
truncated from the _end_, so showing the full path would quickly cause
the name of the actual directory to get lost.