chez: Set `sourceDir` option in config file template
The default source dir is `~/.local/share/chezmoi`, but it can be
changed for any command with the `-S` option.  However, the only way to
avoid specifying it for every command is to set the `sourceDir` option
in the config file.  Fortunately, the value of the `-S` option passed
to the `chezmoi init` command is available in the template data, so the
persistent option can be specified as part of the config file template.