chez: Add bootstrapping of `age` identity file
The easy way to add an `age` identity file would be to make a chezmoi
template that uses the `onepasswordDocument` function.  Unfortunately,
this won't work because `chezmoi` will fail if the config file enables
encryption and specifies an identity file that does not exist yet.

The `init_age.sh` bootstrap script will first make sure that `age`
encryption is installed.  It will then retrieve the identity file from
1Password and write it to the `~/.config/age/me.key.txt` file.  It also
ensures that file and directory permissions are set correctly so that
encryption keys are only readable by the user.

Once the identity file has been bootstrapped into a known location,
then it is safe to enable encryption support in chezmoi.