chez: Enable `age` encryption
If encryption is not enabled, then chezmoi will simply ignore encrypted
files, just as if they were listed in the `.chezmoiignore` file.
However, if encryption is enabled but the specified identity file does
not exist, then chezmoi will exit with an error for every command.  To
prevent this, the config file template only configures encryption if
the expected key file exists.