zsh: Some optimizations for shell initialization
The P10k Instant Prompt hides most of the lag in shell initialization,
so it becomes unnoticeable in most cases.  However, first command lag
is still more significant than it should be.  Some of this lag can be
mitigated with a couple of relatively simple optimizations to reduce the
sub-shell overhead from command substitution.