kitty: Move non-theme settings into main config
Kitty has a theme-selection kitten and a config template which specifies
the color settings that should be included in a theme file.  To make the
configuration compatible with this convention, any extra settings that
are not theme-specific and are not included in the theme template need
to be moved out of the custom theme file and back into the main config.

- <https://sw.kovidgoyal.net/kitty/kittens/themes/>
- <https://raw.githubusercontent.com/kovidgoyal/kitty-themes/master/template.conf>