chez: Use file attachments for SSH private keys
For "SSH Key" items, using the CLI `op read` command to retrieve the
"private key" field will return an exported key file, and not the
original content that was set for the private key.  Unfortunately, the
exported key is unencrypted, and is in a format that is not recognized
by OpenSSH.  Currently, the only workaround is to store the raw private
key file as an attachment, and read the attachment instead of the
dedicated private key field.