zsh: Don't pass dir arg when loading plugin scripts
When a plugin script is sourced from a profile script, the argument
count in the `$#` variable should be zero.  However, if we pass a
directory path to the `_load_plugins` function then the argument count
inside the function will be 1, which could break some plugin scripts.
We only call the function once, so there is no need to pass the path as
an argument.