vim: Update `v` and `vr` shell functions
The `vr` function was not compatible with `nvim`, which does not use the
same `~/.viminfo` format for saving the previous state, so change `vr`
to just trigger the `:History` command from the "fzf.vim" plugin on
startup.  Also move the functions to a sourced file in ".zsh/configs"
to avoid function auto-loading issues.